tests: Fix redundant constants test case to work on x86_64
There currently seems to be a problem with number literals on x86_64,
&hFFFFFFFFFFFFFFFF is a LongInt instead of an Integer, that caused the
I3 constant declarations in this test case to have different dtype.

+++ tests/const/redundant-constants.bas
@@ -9,11 +9,11 @@ const I2 = 123
 const I2 = 123
 
 #ifdef __FB_64BIT__
-	const I3 = &hFFFFFFFFFFFFFFFF
-	const I3 = -1
+	const I3 as integer = &hFFFFFFFFFFFFFFFF
+	const I3 as integer = -1
 #else
-	const I3 = &hFFFFFFFF
-	const I3 = -1
+	const I3 as integer = &hFFFFFFFF
+	const I3 as integer = -1
 #endif
 
 const I4 = &h8000000000000000ull
